[SQL] Move HCatalogTableProvider into its own module
