+++ settings.gradle
@@ -123,6 +123,8 @@ include "beam-sdks-java-extensions-sql-jdbc"
 project(":beam-sdks-java-extensions-sql-jdbc").dir = file("sdks/java/extensions/sql/jdbc")
 include "beam-sdks-java-extensions-sql-shell"
 project(":beam-sdks-java-extensions-sql-shell").dir = file("sdks/java/extensions/sql/shell")
+include "beam-sdks-java-extensions-sql-hcatalog"
+project(":beam-sdks-java-extensions-sql-hcatalog").dir = file("sdks/java/extensions/sql/hcatalog")
 include "beam-sdks-java-fn-execution"
 project(":beam-sdks-java-fn-execution").dir = file("sdks/java/fn-execution")
 include "beam-sdks-java-harness"
